Refactoring vault auth strategies
