+++ vault/app-id_strategy_test.go
@@ -1,7 +1,6 @@
 package vault
 
 import (
-	"net/url"
 	"os"
 	"testing"
 
@@ -38,50 +37,4 @@ func TestNewAppIDAuthStrategy(t *testing.T) {
 	assert.Equal(t, "baz", auth.AppID)
 	assert.Equal(t, "qux", auth.UserID)
 	assert.Equal(t, "quux", auth.Mount)
-}
-
-func TestGetToken_AppIDErrorsGivenNetworkError(t *testing.T) {
-	server, client := setupErrorHTTP()
-	defer server.Close()
-
-	vaultURL, _ := url.Parse("http://vault:8200")
-
-	auth := &AppIDAuthStrategy{"foo", "bar", "app-id", client}
-	_, err := auth.GetToken(vaultURL)
-	assert.Error(t, err)
-}
-
-func TestGetToken_AppIDErrorsGivenHTTPErrorStatus(t *testing.T) {
-	server, client := setupHTTP(500, "application/json; charset=utf-8", `{}`)
-	defer server.Close()
-
-	vaultURL, _ := url.Parse("http://vault:8200")
-
-	auth := &AppIDAuthStrategy{"foo", "bar", "app-id", client}
-	_, err := auth.GetToken(vaultURL)
-	assert.Error(t, err)
-}
-
-func TestGetToken_AppIDErrorsGivenBadJSON(t *testing.T) {
-	server, client := setupHTTP(200, "application/json; charset=utf-8", `{`)
-	defer server.Close()
-
-	vaultURL, _ := url.Parse("http://vault:8200")
-
-	auth := &AppIDAuthStrategy{"foo", "bar", "app-id", client}
-	_, err := auth.GetToken(vaultURL)
-	assert.Error(t, err)
-}
-
-func TestGetToken_AppID(t *testing.T) {
-	server, client := setupHTTP(200, "application/json; charset=utf-8", `{"auth": {"client_token": "baz"}}`)
-	defer server.Close()
-
-	vaultURL, _ := url.Parse("http://vault:8200")
-
-	auth := &AppIDAuthStrategy{"foo", "bar", "app-id", client}
-	token, err := auth.GetToken(vaultURL)
-	assert.NoError(t, err)
-
-	assert.Equal(t, "baz", token)
 }
